Fixed git issues 32(Remove user search searching all users) and 21(moving portlet within same swimlane pops out the assign user popup).

+++ src/main/webapp/WEB-INF/jsp/sprintView.jsp
@@ -400,7 +400,7 @@
 			function showAddUserPopup(elOffset){
 				users_arr = [];
 				$('ul#proj-user-list li').each(function(){
-					$(this).css('border','none');
+					$(this).css('opacity','1');
 					$(this).removeClass('selected');
 				});
 				if($('.popup-story-cont').find('div#pointerEl').hasClass('pointer-rgt')){
@@ -582,7 +582,7 @@
 			        		   					//alert(creatorObj.userName);
 			        		   					refreshStoryPortlet(id.split("st")[1],$(ui.item[0]).closest('ul').attr('id'),creatorObj);
 			        		   				}
-			        		   				if(!($(ui.item[0]).closest('ul').attr('id') == 'notstarted') && !($(ui.item[0]).closest('ul').attr('id') == 'finished')&& !($(ui.item[0]).closest('section').hasClass('left'))){
+			        		   				if((ui.sender != null) && !($(ui.item[0]).closest('ul').attr('id') == 'notstarted') && !($(ui.item[0]).closest('ul').attr('id') == 'finished')&& !($(ui.item[0]).closest('section').hasClass('left'))){
 			        		   					showAddUserPopup(elOffset);
 			        		   					removeUserFromStoryInStage(id.split("st")[1],$(ui.item[0]).closest('ul').attr('id'));
 			        		   					var existing_user_arr = [];
@@ -739,7 +739,6 @@
 			function addUser(userDetails){
 				var stat = false;
 				var post_data = 'userid='+userDetails.username+'&projectId=<%= projectId%>';
-				console.log(userDetails);
 				$.ajax({
 					url: '/scrumr/api/v1/users/create',
 					type: 'POST',
@@ -889,7 +888,6 @@
 	        					$('.stAddmore').hide();
 	        				}
 	        				var people = '';
-	        				console.log(users);
 	        				if(users && users.length > 0){
 	        					var user;
 	        					for(var i=0;i < users.length; i++){
@@ -911,11 +909,11 @@
 			$('ul#proj-user-list li').live("click",function(){
 				if($(this).hasClass("selected")){
 					$(this).removeClass("selected");
-					$(this).css('border','none');
+					$(this).css('opacity','1');
 					users_arr.splice($(this).attr('id').indexOf(),1);
 				}else{
 					$(this).addClass("selected");
-					$(this).css('border','1px solid blue');
+					$(this).css('opacity','0.2');
 					users_arr.push($(this).attr('id'));
 				}
 			});
@@ -924,10 +922,10 @@
 				if(!isAppend){
 					$('.popup-proj-cont').show();
 					$('.popup-proj-cont .c-box-content .user-loading').show();
-					if(addUserScroll){
+					 if(addUserScroll){
 						var api = $('.popup-proj-cont .c-box-content ul').data('jsp');
-						api.destroy();
-					}
+						if(api)api.destroy();
+					} 
 					$('.popup-proj-cont .c-box-content ul').hide();
 				}
 				var post_data = "source=DATABASE&index="+startIndex+"&count=40";
@@ -951,7 +949,6 @@
 								}
 							//	var users_html = "<ul>";
 								var users_html="";
-								console.log(records);
 								var apiVersion = records.success.headers["api-version"];
 								for(var i=0;i<total_users.length;i++){
 									if(!total_users[i].avatar){
@@ -965,6 +962,7 @@
 									$('.popup-proj-cont .c-box-content .jspContainer .jspPane').append(users_html);	
 								}else{
 									$('.popup-proj-cont .c-box-content ul').show();
+									$('.popup-proj-cont .c-box-cont input').attr('id',"searchUser");
 									$('.popup-proj-cont .c-box-content ul').html(users_html);
 								}
 								var atBottom = false;
@@ -1051,20 +1049,23 @@
             }); 
 			
 			$("#removePeople").live('click', function(){
+					 if(addUserScroll){
+						var api = $('.popup-proj-cont .c-box-content ul').data('jsp');
+						if(api)api.destroy();
+					} 
 					var records = users;
-					//var users_html = "<ul>";
+					$('.popup-proj-cont').hide();
 					var users_html="";
 					for(var i=0;i<records.length;i++){
 						if(records[i].username != creator){
 							users_html += '<li><img src="'+records[i].avatarurl+'"/></div><div class="details"><label class="name">'+records[i].fullname+'</label><a class="email">'+records[i].emailid+'</a></div><div style="float:left;" class="removeUser float-rgt disable" id="'+records[i].username+'"></div></li>';
 						}
 					}
-					//users_html += "</ul>";
 					$('.popup-proj-cont .c-box-head').html("Remove people from the project");
+					$('.popup-proj-cont .c-box-content input').attr('id',"searchRmvUser");
 					$('.popup-proj-cont .c-box-content ul').html(users_html);
 					$('.popup-proj-cont').show();
-					$('.popup-proj-cont .c-box-content ul').jScrollPane();
-					//$("#userList-cont").html(users_html);
+					addUserScroll = $('.popup-proj-cont .c-box-content ul').jScrollPane();
 				
 					$(".removeUser").live('click',function(){
 						$(this).css('background', 'url("themes/images/ajax-loader.gif") no-repeat');
@@ -1253,26 +1254,49 @@
     			var selector = $('#storyList').find('ul.story');
     			//query = query.replace(/ /gi, '|'); //add OR for regex query  
 
-    			$(selector).children('li').each(
+    			$(selector).find('li').each(
     					function() {
     						($(this).find('p').text()
     								.search(new RegExp(query, "i")) < 0) ? $(this)
     								.hide() : $(this).show();
     					});
     		});
+        	
+        	$("#searchRmvUser").live('keyup',function(event) {
+    			if (event.which == 13) {
+    				event.preventDefault();
+    			}
+    			var el = $(this);
+    			el.next().css('background','url("themes/images/ajax-loader.gif") no-repeat');
+    			var query=$('#searchRmvUser').val();
+    			var selector = $('.popup-proj-cont .c-box .c-box-content').find('ul');
+    			//query = query.replace(/ /gi, '|'); //add OR for regex query  
+
+    			$(selector).find('li').each(
+    					function() {
+    						($(this).find('label').text()
+    								.search(new RegExp(query, "i")) < 0) ? $(this)
+    								.hide() : $(this).show();
+    					});
+    			el.next().addClass('close-Rmvsearch').css('background','url("themes/images/close.png") no-repeat');
+				$('.close-Rmvsearch').live('click',function(){
+					el.val("");
+	        		el.next().removeClass('close-Rmvsearch').css('background','url("themes/images/search.jpg") no-repeat');
+	        		 $(selector).find('li').each(function(){
+	        			$(this).show();
+	        		}); 
+    			});
+        	});
     		
     		$("#searchUser").live("keyup",function(event) {
     			if (event.which == 13 ||event.which == 8) {
     				event.preventDefault();
     			}
-    			//$(this).closest('.popup-proj-cont').find('.user-loading').show();
-    			//$(this).closest('.popup-proj-cont').find('#total-user-list').hide();
-    			var query=$('#searchUser').val();
-    			//var selector = $('.popup-proj-cont .c-box .c-box-content').find('ul');
-    			var post_data = "sortType="+query+"&showTotalCount=false&startIndex=1&count=20";
     			var el = $(this);
     			el.next().css('background','url("themes/images/ajax-loader.gif") no-repeat');
     			setTimeout(function(){
+    				var query=$('#searchUser').val();
+        			var post_data = "sortType="+query+"&showTotalCount=false&startIndex=1&count=20";
     				$.ajax({
 						url : '/scrumr/api/v1/users/searchqontext/',
 						type : 'POST',
@@ -1397,12 +1421,12 @@
 		       	$(this).parent().css('background-color',"#F6EEE1");
 				$(".sprintview").css('color',"gray");
 				$(".sprintview").parent().css('background-color',"#FFFFFF");
-		       	populateSprints();
 				$("#sprint-view").hide();
 				$("#project-view").show();
 				$(".duration-hd").find('label').show();
 				$(".duration-hd").find('#pageCtrls').show();
 				$(".duration-hd").find('ul').hide();
+				populateSprints();
 				viewStoryFancyBox();
 			});
 			
@@ -1413,12 +1437,12 @@
 				$(this).parent().css('background-color',"#F6EEE1");
 		       	$(".projectview").css('color',"gray");
 		       	$(".projectview").parent().css('background-color',"#FFFFFF");
-				populateSprintStories(current_sprint);
 				$("#sprint-view").show();
 				$("#project-view").hide();
 				$(".duration-hd").find('label').hide();
 				$(".duration-hd").find('#pageCtrls').hide();
 				$(".duration-hd").find('ul').show();
+				populateSprintStories(current_sprint);
 				viewStoryFancyBox();
 			});
